+++ internal/validation/validation.go
@@ -1,34 +1,36 @@
 package validation
 
 import (
-	"fmt"
+	"log"
 	"net/http"
 	"strings"
 
 	c "github.com/sanek1/metrics-collector/internal/config"
 )
 
+var logger *log.Logger
+
+func init() {
+	logger = log.New(log.Writer(), "VALIDATION: ", log.Ldate|log.Ltime|log.Lshortfile)
+}
+
 func Validation(next http.Handler) func(http.ResponseWriter, *http.Request) {
 	return http.HandlerFunc(func(rw http.ResponseWriter, r *http.Request) {
-		message := "-- start validation --"
-		printValidationMessage(message)
+		logValidationMessage("-- start validation --")
 		rw.Header().Set("Content-Type", "application/json")
-
-		// check path
 		splitedPath := strings.Split(r.URL.Path, "/")
 		if len(splitedPath) < c.MinPathLen {
-			message = "invalid path"
+			message := "invalid path"
 
-			printValidationMessage(message)
+			logValidationMessage(message)
 			http.Error(rw, message, http.StatusNotFound)
 			return
 		}
-
 		next.ServeHTTP(rw, r)
-		printValidationMessage("--validation completed successfully--")
+		logValidationMessage("--validation completed successfully--")
 	})
 }
 
-func printValidationMessage(message string) {
-	fmt.Println(message)
+func logValidationMessage(message string) {
+	logger.Println(message)
 }
